pyproject.toml
[project]
name = "hallm"
version = "0.1.0"
description = "LLM-powered assistant with MCP server and CLI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aioboto3>=13",
    "asyncpg>=0.30",
    "environs>=14.0",
    "fastmcp>=2.0",
    "httpx>=0.28.1",
    "ipdb>=0.13",
    "opentelemetry-api>=1.27",
    "opentelemetry-exporter-otlp>=1.27",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-sdk>=1.27",
    "redis>=5",
    "sentry-sdk>=2",
    "slugify>=0.0.1",
    "tortoise-orm[asyncpg]>=0.22",
    "typer>=0.25.1",
    "validators>=0.35.0",
]

[project.scripts]
hallm = "hallm.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hallm"]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pre-commit>=4.6.0",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.11",
    "ty>=0.0.1a8",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
src = ["."]
line-length = 100
target-version = "py314"
exclude = ["hallm/db/migrations"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hallm"]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Ty
# ---------------------------------------------------------------------------
[tool.ty.src]
include = ["."]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short -q --cov=hallm --cov-report=term-missing --cov-report=xml"
markers = [
    "integration: marks tests as integration tests that call real APIs (deselect with '-m not integration')",
]

[tool.coverage.run]
source = ["hallm"]
branch = true
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 98
skip_covered = false
precision = 2
show_missing = true

# ---------------------------------------------------------------------------
# Tortoise ORM / migrations
# ---------------------------------------------------------------------------
[tool.tortoise]
tortoise_orm = "hallm.db.TORTOISE_ORM"
